move the SA constants to SaProcessor.cpp
