+++ plugins/SpectrumAnalyzer/CMakeLists.txt
@@ -4,5 +4,4 @@ include_directories(SYSTEM ${FFTW3F_INCLUDE_DIRS})
 LINK_LIBRARIES(${FFTW3F_LIBRARIES})
 
 BUILD_PLUGIN(analyzer Analyzer.cpp SaProcessor.cpp SaControls.cpp SaControlsDialog.cpp SaSpectrumView.cpp SaWaterfallView.cpp
-MOCFILES SaProcessor.h SaConstants.h SaControls.h SaControlsDialog.h SaSpectrumView.h SaWaterfallView.h DataprocLauncher.h 
-EMBEDDED_RESOURCES *.svg logo.png)
+MOCFILES SaProcessor.h SaControls.h SaControlsDialog.h SaSpectrumView.h SaWaterfallView.h DataprocLauncher.h EMBEDDED_RESOURCES *.svg logo.png)
